Refactor config routes, menu for all apps
- Use named export
- Consolidate the types in a /config/types folder
- Simplify the types where possible; did not manage to get rid of the per route-types, though; but they are simplified
- Add types where missing
