+++ src/config/aachen/index.ts
@@ -2,10 +2,9 @@ import { map as defaultMapConfig } from '~/config/default/apps';
 import defaultBase from '~/config/default/base';
 import defaultColors from '~/config/default/colors';
 import { RegionConfig } from '~/types';
-
-import menu from './menu';
-import routes from './routes';
-import staticpages from './staticpages';
+import { menu } from './menu';
+import { routes } from './routes';
+import { staticpages } from './staticpages';
 
 const apiEndpoints = {
   local: 'http://localhost:8000/api',
