+++ src/config/default/index.ts
@@ -1,11 +1,10 @@
 import { DefaultConfig } from '~/types';
-
 import { map } from './apps';
 import baseConfig from './base';
 import colors from './colors';
-import menu from './menu';
-import routes from './routes';
-import staticpages from './staticpages';
+import { menu } from './menu';
+import { routes } from './routes';
+import { staticpages } from './staticpages';
 
 const defaultConfig: DefaultConfig = {
   apps: {
