+++ src/config/eichwalde/menu.ts
@@ -1,8 +1,9 @@
-import PlanningIcon from '~/images/planning-icons/planung.svg';
 import AboutIcon from '~/images/info.svg';
+import PlanningIcon from '~/images/planning-icons/planung.svg';
+import { ConfigMenu } from '../types/ConfigMenu';
 import { routes } from './routes';
 
-export const menu = {
+export const menu: ConfigMenu = {
   size: 325,
   profileLabel: 'Zum Profil',
   loginLabel: 'Login',
