+++ src/config/eichwalde/routes.ts
@@ -1,7 +1,15 @@
-import defaultRoutes from '../default/routes';
+import {
+  ConfigMandatoryRoutes,
+  mandatoryRoutes,
+} from '../default/mandatoryRoutes';
+import { ConfigRoutePath } from '../types/ConfigRoute';
 
-export const routes = {
-  ...defaultRoutes,
+type Props = ConfigMandatoryRoutes & {
+  landing: ConfigRoutePath;
+  cycleNetworkDevelopment: ConfigRoutePath;
+};
+export const routes: Props = {
+  ...mandatoryRoutes,
   landing: '/forschungsprojekt',
   cycleNetworkDevelopment: '/radnetzentwicklung',
 };
