+++ src/config/types/ConfigRoute.ts
@@ -1,0 +1,5 @@
+export type ConfigRoutePath = `/${string}`;
+
+export type ConfigRoutePage = {
+  [key: string]: ConfigRoutePath;
+};
