+++ src/config/types/ConfigStaticPages.ts
@@ -1,0 +1,6 @@
+import { ConfigRoutePath } from './ConfigRoute';
+
+export type ConfigStaticpages = {
+  key: string;
+  route: ConfigRoutePath;
+};
